+++ core/components/s3extended/lexicon/en/default.inc.php
@@ -7,5 +7,6 @@
 //           default.inc.php - default lexicon              //
 //                                                         ///
 //////////////////////////////////////////////////////////////
+$_lang['s3extended.code'] = 's3e';
 $_lang['s3extended.name'] = 'Amazon S3 Extended Über Edition';
 $_lang['s3extended.description'] = 'Amazon S3 extended is a fork of the core ModX Amazon S3 functionality. It adds resize before upload and much more!';